refactor(l5): move nats-bridge from src/ into server/application (issue 864)
L5 of _plans/server-hardening-ladder.md: cut server/ free of src/. The NATS
context-pack bridge was a single 971-line module in src/; it moves to
server/application/ and splits along the seams it already had.

Moved and split:

  src/nats-bridge.ts -> server/application/nats-bridge.ts        (357 lines)
    health, the start loop, and the reply seam (respondWithinBrokerFigure
    and the #549 undeliverable-reply paths).
  -> server/application/nats-bridge-types.ts                     (91 lines)
    the wire/driver/runtime interfaces and the module constants, shared by
    every sibling so none of them imports another for a type alone.
  -> server/application/nats-bridge-envelope.ts                  (186 lines)
    envelope parse/build, the error envelopes, the redacting error
    classifiers and the three log emitters.
  -> server/application/nats-bridge-handler.ts                   (310 lines)
    handleNatsContextPackMessage, lane binding, and header auth.
  -> server/application/nats-bridge-driver.ts                    (253 lines)
    createNatsJsDriver and the resubscribe loop.

Every sibling is under 500 code lines, every function under 100, complexity
under 10, at most 4 parameters, and none reads process.env.

No behavior change. The NATS subjects, envelope shapes, error strings, error
codes and resubscribe delays are byte-identical to the pre-move file. Three
extractions were forced by the server/ lint standard and are mechanical:

  - authFromHeaders' three-casing header read is hoisted into
    rawAuthorizationHeader (same "??" chain, same order).
  - the resubscribe loop's fixed collaborators (connection, subject, handler,
    health) move into the ResubscribeState object as readonly fields, and the
    drain try/catch becomes drainOnce, which returns false on the throw path
    exactly as the old inline processedMessage did.
  - respondWithinBrokerFigure's two halves become tryPublishReply and
    publishUndeliverableReply, preserving all four reply outcomes.

Importers:

  server/application/nats.ts            ../../src/nats-bridge.ts -> ./nats-bridge.ts
  server/application/operator-doctor.ts ../../src/nats-bridge.ts -> ./nats-bridge.ts (type, and its re-export)

src/nats-bridge.ts is an M2 shim: every exported signature already met the
server standard (options objects or two parameters), so no adapter is needed.
Its importers -- src/index.ts, src/nats-worker.ts, src/tools/index.ts,
src/server.test.ts, src/nats-bridge.test.ts, src/nats-worker.test.ts,
src/tools/__tests__/get-contract.test.ts and the three scripts/run-nats-worker
tests -- keep their import lines untouched and retire with src/ at L6.

Imports inside the moved files: auth, background-tracing and nats-runtime
point at their server/application/ homes; logger stays on src/logger.ts, and
tools/agent-context-pack.ts, tools/index.ts (type-only) and types.ts stay on
their src/ paths per M3, for later lanes.

Tests are untouched (M1 exception): src/nats-bridge.test.ts and
src/nats-worker.test.ts keep testing the moved code through the shim.

src/ runtime closure under server/main.ts: 14 -> 13.

Verified:
  bunx tsc --noEmit -> exit 0
  oxlint --deny-warnings on all 8 staged paths -> exit 0
  bun run test:isolated (8 named files) -> 115 tests, 0 fail, same as origin/main
  bun contracts/check-parity.ts -> exit 0
  bash scripts/done-means/864-moved-out-of-src.sh -> PASS, judged=41 shims, 12 adapters
